requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.2.0
WeasyPrint==60.2
